refactor: extract entrapment metrics module
